requirements.txt
Pillow==9.2.0
scikit_learn==1.1.1
streamlit==1.11.0
